+++ game_patch/os/os.cpp
@@ -12,10 +12,7 @@
 #include "../multi/multi.h"
 #include "os.h"
 #include "win32_console.h"
-
 #include <timeapi.h>
-
-const char* get_win_msg_name(UINT msg);
 
 FunHook<void()> os_poll_hook{
     0x00524B60,
@@ -75,7 +72,7 @@ LRESULT WINAPI wnd_proc(HWND wnd_handle, UINT msg, WPARAM w_param, LPARAM l_para
         }
 
         rf::is_main_wnd_active = w_param;
-        return 0; //DefWindowProcA(wnd_handle, msg, w_param, l_param);
+        return 0;
 
     case WM_WINDOWPOSCHANGING:
         if (is_headless_mode() && l_param) {
@@ -240,6 +237,17 @@ bool awpgen_requested_from_raw_cmdline()
 bool headless_requested_from_raw_cmdline()
 {
     return headless_bot_requested_from_raw_cmdline() || awpgen_requested_from_raw_cmdline();
+}
+
+void wnd_set_flash(const HWND hwnd) {
+    FLASHWINFO flash{
+        .cbSize = sizeof(flash),
+        .hwnd = hwnd,
+        .dwFlags = FLASHW_TRAY,
+        .uCount = 3ul,
+        .dwTimeout = 0
+    };
+    FlashWindowEx(&flash);
 }
 
 static FunHook<void()> os_close_hook{