+++ docs/CHANGELOG.md
@@ -21,6 +21,8 @@ Version 1.4.0 (Lupin): Not yet released
 - Make clock clutter objects correctly display the current local real world time
 
 [@is-this-c](https://github.com/is-this-c)
+- Add `mp_join_flash` console command to flash your window upon player joins, if your window is out of focus
+- Rename `mp_notifyonjoin` console command to `mp_join_beep`
 - Disable `Refresh Selected` in the server browser, only if `Refresh Selected` was pressed
 - Never disable `Add Server` in the server browser
 - Add `NOT IN ROUND`, `IDLE`, and `SPECTATOR` to the spectate UI
@@ -58,14 +60,14 @@ Version 1.3.0 (Bakeapple): Released Apr-22-2026
   - Support `Alpha` field in `Decal` objects placed in version >= 304 levels
   - Support movers, meshes, and mesh pixel lighting in skyboxes
   - Support gas regions
-- Expanded destruction capabilities available to developers
+- Expand destruction capabilities available to developers
   - `Brush-based geomod` switch added to level properties; if true, use level hardness for geoable brushes (RF2-style)
   - Geo regions allow traditional world-based geomod to be used even when brush-based switch is true
   - `Is Geoable` flag added to brush properties
   - Support for rock, wood, cement, metal, and ice breakable detail brushes
   - Add dynamic debris generation from breakable detail brushes
   - `No Debris` flag added to brush properties for breakable detail brushes
-- Added new object types
+- Add new object types
   - `Mesh` for configuring custom static, skeletal, or animated meshes in levels
   - `Corona` for configuring custom glare effects in levels
   - `Note` for leaving important information in levels (editor only)